fix to backstack/stores remote events when app is opened
