+++ app/src/main/java/org/cpen321/discovr/MainActivity.java
@@ -27,7 +27,11 @@ import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.widget.Button;
 import android.widget.Toast;
+
+import com.loopj.android.http.AsyncHttpClient;
+import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.mapbox.mapboxsdk.MapboxAccountManager;
 import android.view.View;
 import com.mapbox.mapboxsdk.MapboxAccountManager;
@@ -55,12 +59,19 @@ import com.mapbox.services.directions.v5.models.DirectionsRoute;
 import com.mapbox.services.geocoding.v5.MapboxGeocoding;
 import com.mapbox.services.geocoding.v5.models.CarmenFeature;
 import com.mapbox.services.geocoding.v5.models.GeocodingResponse;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Locale;
 
+import cz.msebera.android.httpclient.Header;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
@@ -89,12 +100,56 @@ public class MainActivity extends AppCompatActivity
     //Building information JSON inputstream for searching
     InputStream is;
 
+    private List<EventInfo> AllEventsList = new ArrayList<EventInfo>();
     private static final int REQUEST_ALL_MAPBOX_PERMISSIONS = 3211;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {// Get the SearchView and set the searchable configuration
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        // Inflate the layout for this fragment
+
+        AsyncHttpClient client = new AsyncHttpClient();
+        client.get("http://discovrweb.azurewebsites.net/api/Events", new AsyncHttpResponseHandler() {
+            @Override
+            public void onStart() {
+                // called before request is started
+            }
+
+            @Override
+            public void onSuccess(int statusCode, Header[] headers, byte[] response) {
+                String r = new String(response);
+                try {
+                    JSONArray json = new JSONArray(r);
+                    for(int i = 0; i < json.length(); i++){
+                        JSONObject o = json.getJSONObject(i);
+                        AllEventsList.add(new EventInfo(o.getInt("Id"),
+                                o.getString("Name"),
+                                o.getString("Host"),
+                                o.getString("Location"),
+                                o.getString("StartTime"),
+                                o.getString("EndTime"),
+                                "",
+                                o.getString("Description")));
+                    }
+                }
+                catch (JSONException e){
+                    throw new RuntimeException(e);
+                }
+                System.out.println(r);
+            }
+
+            @Override
+            public void onFailure(int statusCode, Header[] headers, byte[] errorResponse, Throwable e) {
+                // called when response HTTP status is "4XX" (eg. 401, 403, 404)
+                System.out.println(":(");
+            }
+            @Override
+            public void onRetry(int retryNo) {
+                // called when request is retried
+            }
+        });
 
         // TODO: Refactor permission code
         String[] permissions = {
@@ -474,7 +529,7 @@ public class MainActivity extends AppCompatActivity
             }
         }
         Log.d("backstack", "Destroying backstack of size: " + fm.getBackStackEntryCount());
-        //fm.popBackStackImmediate(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+        fm.popBackStackImmediate(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
         Log.d("backstack", "Backstack size: " + fm.getBackStackEntryCount());
 
         //Adds a fragment to the container and changes the toolbar title correspondingly
@@ -502,5 +557,9 @@ public class MainActivity extends AppCompatActivity
         }
 
         ft.commit();
+    }
+
+    List<EventInfo> getAllEvents(){
+        return this.AllEventsList;
     }
 }
